fixed problems with db regeneration

+++ plugins/pybox/shortcode-make-databases.php
@@ -27,7 +27,11 @@ function makedb($content, $options) {
 
   function cmp($l1, $l2) {
     $c = strcmp($l1['lang'], $l2['lang']);
-    if ($c != 0) return $c;
+    if ($c != 0) {
+      if ($l1['lang']=='en') return -1; // put english first
+      if ($l2['lang']=='en') return 1;
+      return $c;
+    }
     if ($l1['number'] == NULL ^ $l2['number'] == NULL) {
       return $l1['number'] == NULL ? 1 : -1;
     }
@@ -39,52 +43,51 @@ function makedb($content, $options) {
 
   pyboxlog('[makedb] Rebuilding problem and lesson db', TRUE);
 
-  global $wpdb;
+  global $wpdb, $SKIP_DB_REBUILD;
+  $SKIP_DB_REBUILD = false; // set to true to be diagnostic
 
   $lesson_table_name = $wpdb->prefix . "pb_lessons";
   $problem_table_name = $wpdb->prefix . "pb_problems";
-  $wpdb->query("TRUNCATE TABLE $problem_table_name;");
-  $wpdb->query("TRUNCATE TABLE $lesson_table_name;");
+
+  if (!$GLOBALS['SKIP_DB_REBUILD']) {
+    echo 'Truncating tables ';
+    $wpdb->query("TRUNCATE TABLE $problem_table_name;");
+    $wpdb->query("TRUNCATE TABLE $lesson_table_name;");
+  }
 
   $currlang = 'xx';
   $i = -1;
   $gets = array();
   foreach ($lessons as $l) {
+    echo '<pre>';
     if ($currlang != $l['lang']) {
       $currlang = $l['lang'];
       $i = 0;
     }
-    $l['ordering'] = $i;    
-    echo "<div id='x$currlang$i'>";
-    if ($l['number'] != NULL)
-      echo ($wpdb->insert($lesson_table_name, $l)!=1?'insert bad':'insert ok');
-    echo ' ' . json_encode($l) . "</div>\n";
-    $url = get_page_link($l['id']);
-    $url = str_replace('/dev', '', $url);
-    $index = $l['number'] == NULL?-1:$i;
-    $gets[] = array('url'=>$url, 'makeproblemdb'=>$index, 'lessonlang'=>$currlang, 'lessonnumber'=>$l['number'],
-		    'id'=>'#x'.$currlang .$i);
-    $i++;
-  }
-  echo '<script type="text/javascript">
-var gets = '.json_encode($gets).';
-foo = function(a, asynch) {
-   jQuery.ajax({type:"GET",
-               url:a["url"],
-               async:asynch,
-               data:jQuery.param({"makeproblemdb":a["makeproblemdb"],
-                                  "lessonlang":a["lessonlang"],
-                                  "lessonnumber":a["lessonnumber"]}),
-               success:function(data){jQuery(a["id"]).append("<br/>retrieved, "+data.length+" bytes</br>");}});
-}
-for (var i=0; i<gets.length; i++)
-  if (gets[i]["lessonlang"]=="en")
-    foo(gets[i], false);  
-for (var i=0; i<gets.length; i++)
-  if (gets[i]["lessonlang"]!="en")
-    foo(gets[i], true);  
-</script>';
+    if ($l['number'] == NULL) 
+      $index = -1;
+    else {
+      $index = $i;
+      $i++;
+    }
+    $l['ordering'] = $index;
 
+    if ($l['number'] != NULL) {
+      echo 'About to insert lesson: ' . rowSummary($l); //. json_encode($l);
+      if (!$GLOBALS['SKIP_DB_REBUILD']) 
+	echo ($wpdb->insert($lesson_table_name, $l)!=1?'<br>insert bad':' insert ok');
+    }
+
+    global $lesson_reg_info, $pyRenderCount, $post;
+    $lesson_reg_info = array('index'=>$index, 'lang'=>$currlang, 'fullnumber'=>$l['number'],
+			   'url'=>str_replace('/dev','', get_page_link($l['id'])), 'id'=>$l['id']);
+
+    // render! the following line is not just cosmetic as it registers the problems.
+    echo '<br>Snippet: ' . htmlspecialchars(substr(do_shortcode(get_page($l['id'])->post_content), 0, 50));
+    $pyRenderCount = 0; // to get problem links working
+    echo '</pre>';
+  }
+  $lesson_reg_info = 1;
 }
 
 // end of file
